files relocated according to reviewer suggestion in #120
