+++ relecov_dashboard/utils/generic_graphic_data.py
