+++ relecov_dashboard/utils/met_sample_preprocessing.py
@@ -3,9 +3,9 @@ import pandas
 
 # Local imports
 import relecov_core.utils.rest_api_handling
-import relecov_dashboard.utils.generic.generic_graphic_data
-import relecov_dashboard.utils.generic.graphics.plotly
-import relecov_dashboard.utils.generic.generic_process_data
+import relecov_dashboard.utils.generic_graphic_data
+import relecov_dashboard.utils.plotly
+import relecov_dashboard.utils.generic_process_data
 
 
 def sample_processing_graphics():
